# packages/api-client/openapi/v1.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Shellwright control plane",
    "description": "Organisations, apps, and immutable configuration versions.\n\nErrors are RFC 9457 problem documents. The `type` URI and the `code` extension are stable; the `title` and `detail` are for people and may be reworded.",
    "version": "v1"
  },
  "paths": {
...
    "/v1/orgs/{orgId}/tokens": {
      "get": {
        "tags": ["API tokens"],
        "summary": "List an organisation's API tokens.",
        "parameters": [
          {
            "name": "orgId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApiTokenResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["API tokens"],
        "summary": "Mint an API token and show the secret once.",
        "parameters": [
          {
            "name": "orgId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApiTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedApiTokenResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/orgs/{orgId}/tokens/{tokenId}": {
      "delete": {
        "tags": ["API tokens"],
        "summary": "Revoke an API token.",
        "parameters": [
          {
            "name": "orgId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "tokenId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/v1/workspaces/{workspaceId}/apps": {
      "get": {
        "tags": ["Apps"],
        "summary": "List a workspace's apps.",
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AppResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Apps"],
        "summary": "Create an app and seed its first configuration.",
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAppRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/apps/{appId}": {
      "get": {
        "tags": ["Apps"],
        "summary": "Describe an app.",
        "parameters": [
          {
            "name": "appId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiTokenResponse": {
        "required": ["id", "name", "prefix", "role", "workspaceId", "createdAt", "lastUsedAt"],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Identifier.",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "description": "Label."
          },
          "prefix": {
            "type": "string",
            "description": "Leading characters, enough to tell two tokens apart."
          },
          "role": {
            "type": "string",
            "description": "Ceiling on what it may do."
          },
          "workspaceId": {
            "type": "string",
            "description": "Workspace it is confined to, if any.",
            "format": "uuid",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "description": "When it was minted.",
            "format": "date-time"
          },
          "lastUsedAt": {
            "type": "string",
            "description": "Coarse last use.",
            "format": "date-time",
            "nullable": true
          }
        },
        "description": "A token as listed. Never contains the secret."
      },
      "AppResponse": {
        "required": ["id", "workspaceId", "name", "bundleId", "currentConfigVersionId", "createdAt"],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Identifier.",
            "format": "uuid"
          },
          "workspaceId": {
            "type": "string",
            "description": "Owning workspace.",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "description": "Display name."
          },
          "bundleId": {
            "type": "string",
            "description": "Reverse-DNS identifier."
          },
          "currentConfigVersionId": {
            "type": "string",
            "description": "The version considered live, if any.",
            "format": "uuid",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "description": "Creation timestamp.",
            "format": "date-time"
          }
        },
        "description": "An app as the API reports it."
      },
...
      "CreateApiTokenRequest": {
        "required": ["name"],
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Human-readable label."
          },
          "role": {
            "$ref": "#/components/schemas/OrgRole"
          },
          "workspaceId": {
            "type": "string",
            "description": "Workspace to confine it to, or null for the organisation.",
            "format": "uuid",
            "nullable": true
          }
        },
        "description": "Mint an API token."
      },
      "CreateAppRequest": {
        "required": ["name", "bundleId", "initialUrl"],
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Display name, also the name under the icon."
          },
          "bundleId": {
            "type": "string",
            "description": "Reverse-DNS identifier, shared by both platforms."
          },
          "initialUrl": {
            "type": "string",
            "description": "The first page the shell loads."
          }
        },
        "description": "Create an app."
      },
      "CreatedApiTokenResponse": {
        "required": ["token", "details"],
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "⚠️ The only time the secret is ever available."
          },
          "details": {
            "$ref": "#/components/schemas/ApiTokenResponse"
          }
        },
        "description": "A token as returned once, at creation."
      },
...
      "OrgRole": {
        "enum": ["Viewer", "Developer", "Admin", "Owner"],
        "description": "Membership roles, ordered from least to most privileged.",
        "default": "Developer"
      },
...
    }
  },
...
}